requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pywhatkit==5.4.3
requests==2.31.0
plyer==2.1.0
pycaw==20230407
comtypes==1.2.0
python-dotenv==1.0.0 